app/backend/requirements.txt
Django==2.2.10
pytz==2019.2
sqlparse==0.3.0
djangorestframework==3.11.0
pillow==6.2.1
django-cors-headers==3.2.0
django-rest-auth==0.9.3
django-allauth==0.37.1